Revert "Reporter: Fix YAML output in TAP reporter (#110)"
This turns too many values into quoted JSON strings, which
makes the output difficult to read.

I am reverting this for 1.x-stable to allow other changes to be
released and adopted by consumers, such as QUnit.

I will improve on the solution in a different way in master
for js-reporters 2.0.

Ref https://github.com/js-reporters/js-reporters/issues/109.