Reporter: TapReporter align actual with expected
When comparing two strings, it is significantly easier to spot the
difference when they are aligned. This change pads the label for
`actual` with a couple of spaces so that the actual value is printed at
the same indentation as the expected value. Previously the expected
value was right shifted by 2 characters...

Closes https://github.com/js-reporters/js-reporters/pull/107.