Build: Update commitplease dev dependency to support nohook
The 'nohook' option didn't work properly on older versions.
Also enable it by default for now so that 'npm ci' works
out-of-the-box without needing to grant write access to the
.git directory in my dev container, which is a a security risk.

Also commit 'package-lock.json' to make builds faster and more
deterministic, reducing the chances that some indirect update
might cause a build issue and then for no-one to know which version
of what last worked correctly.