[BREAKING CHANGE] Build: Use ES6 plainly (remove Rollup and Babel)
The current stack is imho too complicated to maintain and not
worth the benefit. I'm expecting this to have fairly little impact
since apart from browserstack and qunit, there are no dependants
of this package on npm.

To allow me to effectively pick up maintenance for js-reporters,
optimise it for use in Node.js

For now, use in a browser will require downstream to bundle this
module as a library with their payload. I'm open to reconsidering
this in the future if there's demand for it, e.g. by documenting
how to create a simple ES5-compatible bundle and perhaps by
offering a premade es5/commonjs artefact for download.