MochaAdapter: Fix TypeErorr due to arrow-function reporter
Follows-up 370827391243fc, which revealed that the only reason
this code previously worked is that we compiled it to ES5 where
the arrow function became a normal function.

The CI build did not catch this because the integration test
for QUnit/Jasmine/Mocha is passing for QUnit and Jasmine and then
aborts silently without exit code due to a different bug, so it
has not actually been testing this code for a while, unfortunately.

I will address that in a separate commit soon.