Test: Remove redundant Mocha stdout override
Per d7d03961cfdd, the Mocha integration test isn't working currently.

This presumably existed because Mocha either couldn't handle
nested state well, or was causing noise in the output from the
nested test.

However in practice what happened is that it never restored,
and hid the rest of our own top-level test, and given the test
was failing it also subsequently caused there to not be any
non-zero exit code for CI to fail the build.

It seems to be redundant, because simply removing this means
errors are coming through properly, and there does not appear
to be any noise in the output from the nested test.