Test: Convert test suite from Mocha to QUnit
It even found a bug, due to QUnit's deepEqual being stricter it
seems. The 'assertions' arrays of TestEnd objects contains
Assertion objects, but our test fixture is saying it expects
an array of plain objects. This discrepancy was previously
tolerated.

I've switched the offending assertion to use propEqual() for now,
as I don't to touch any actual source code or expected test values
in this refactor.