Test: Install integration versions in a tmpdir and add logging
Make the "npm run test-versions" script (integration/adapters-run.js)
more robust by installing packages in a temporary directory
entirely separate from the repository's working directory.

Also capture the output of the test runs in a file so that one
can easily inspect and find out why a version is failing. This
can then be used to report issues upstream (if it's a new version),
or to fix something on our end of it is genuine, or to document
it in failing-versions.js if it is a known issue with a past release.

- qunitjs:
  Make the list of old versions that are failing complete
  and document why they fail.

- jasmine:
  Version 2.5.2 and later failed due to a default reporter killing
  killing the process (again). This needs to be fixed on our side
  by clearing the default reporters. Fixed as part of this commit
  since it didn't need a change to the adapter.

  Version 3.0 and later failed due to randomized order.
  Disable in our test.

- mocha:
  Version 8.0 and later are currently failing due to skipped
  tests not being recognised correctly. To be followed up on
  in a later commit as it looks like that would require a source
  change to the adapter.