Test: Radically simplify integration test reference data
Remove all the computed indirection from the test fixtures.

The expected/reference values should be as plain as possible
so that it is easy to change things, and easy to review and
confirm how things are supposed to work, and to have high
confidence in the results when they arre passing.

Previously, the expected results were complicated and thus
presented their own source of potential bugs and require debugging.

Instead:

- Use object literals for the expected values.
- Use propEqual() for comparing actual with expected.
- Any tolerance or mutation is done in a single place, before
  the tests begin.