Test: Tolerate additional properties on Assertion event data
This was already in the spec, but the test was extra strict,
which would have caused other compliant implementations to
fail our test.

Also add a test for the names of tests and suites before we compare
arbitrary offsets in the result arrays. This makes off-by-one
failures less confusing to debug.

Ref https://github.com/js-reporters/js-reporters/issues/105.